backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
passlib[bcrypt]>=1.7.4
bcrypt<4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
requests
